Start working on more compression of the shorts and making it functional again.

+++ Aurora/BotManager/GenericNpcCharacter.cs
@@ -463,19 +463,19 @@ namespace Aurora.BotManager
 
         }
 
-        public virtual void SendLayerData(float[] map)
+        public virtual void SendLayerData(short[] map)
         {
         }
 
-        public virtual void SendLayerData(int px, int py, float[] map)
+        public virtual void SendLayerData(int px, int py, short[] map)
         {
         }
 
-        public void SendLayerData(int[] x, int[] y, float[] map, TerrainPatch.LayerType type)
+        public void SendLayerData(int[] x, int[] y, short[] map, TerrainPatch.LayerType type)
         {
 
         }
-        public virtual void SendLayerData(int px, int py, float[] map, bool track)
+        public virtual void SendLayerData(int px, int py, short[] map, bool track)
         {
         }
